+++ src/pages/Login.tsx
@@ -1,12 +1,11 @@
 import React, { Dispatch, SetStateAction, useState, useCallback } from "react";
 import { Form, Input, Button } from "antd";
 import { useNavigate } from "react-router-dom";
-import { googleLogin, kakaoLogin, login } from "lib/api/user";
+import { login } from "lib/api/user";
 import styled from "styled-components";
-import GoogleLogin from "react-google-login";
-import KakaoLogin from "react-kakao-login";
 import { useMutation } from "react-query";
 import { UserType } from "interfaces/interfaces";
+import { SocialLogin } from "components";
 import Cookies from "universal-cookie";
 
 const cookies = new Cookies();
@@ -18,8 +17,6 @@ interface LoginType {
 const Login: React.FunctionComponent<LoginType> = ({ setIsLogin }) => {
   const navigate = useNavigate();
   const mutationLogin = useMutation((user: UserType) => login(user));
-  const mutationGoogle = useMutation((token: string) => googleLogin(token));
-  const mutationKakao = useMutation((token: string) => kakaoLogin(token));
 
   const [user, setUser] = useState({ email: "", password: "" });
   const [errorMsg, setErrorMsg] = useState("");
@@ -58,44 +55,6 @@ const Login: React.FunctionComponent<LoginType> = ({ setIsLogin }) => {
         setErrorMsg("로그인에 실패했습니다");
       });
   }, [mutationLogin, setIsLogin, setToken, user]);
-  const _handleGoogleSuccess = useCallback(
-    (token: any) => {
-      mutationGoogle
-        .mutateAsync(token.tokenObj.id_token)
-        .then((res) => {
-          if (res.data.success) {
-            setIsLogin(true);
-            setToken(res.data.data);
-          }
-        })
-        .catch(() => {
-          setErrorMsg("로그인에 실패했습니다");
-        });
-    },
-    [mutationGoogle, setIsLogin, setToken],
-  );
-  const _handleGoogleFailure = useCallback(() => {
-    console.log("Google Login Failure");
-  }, []);
-  const _handleKakaoSuccess = useCallback(
-    (token: any) => {
-      mutationKakao
-        .mutateAsync(token.response.access_token)
-        .then((res) => {
-          if (res.data.success) {
-            setIsLogin(true);
-            setToken(res.data.data);
-          }
-        })
-        .catch(() => {
-          setErrorMsg("로그인에 실패했습니다");
-        });
-    },
-    [mutationKakao, setIsLogin, setToken],
-  );
-  const _handleKakaoFailure = useCallback(() => {
-    console.log("Kakao Login Failure");
-  }, []);
 
   return (
     <Container>
@@ -152,40 +111,12 @@ const Login: React.FunctionComponent<LoginType> = ({ setIsLogin }) => {
               로그인
             </StyledButton>
           </ButtonBox>
-          <SocialButtons>
-            <SocialLoginButton>
-              <GoogleLogin
-                buttonText="    Sign up with Google    "
-                clientId={process.env.REACT_APP_GOOGLE_CLIENT_ID || ""}
-                onSuccess={_handleGoogleSuccess}
-                onFailure={_handleGoogleFailure}
-                disabled={mutationLogin.isLoading}
-              />
-            </SocialLoginButton>
-            <SocialLoginButton>
-              <KakaoLogin
-                token={process.env.REACT_APP_KAKAO_TOKEN || ""}
-                onSuccess={_handleKakaoSuccess}
-                onFail={_handleKakaoFailure}
-                render={({ onClick }) => (
-                  <div
-                    onClick={(e) => {
-                      e.preventDefault();
-                      if (!mutationLogin.isLoading) {
-                        onClick();
-                      }
-                    }}
-                  >
-                    <img
-                      alt="kakao"
-                      src="/kakao_login_medium_wide.png"
-                      width={250}
-                    />
-                  </div>
-                )}
-              />
-            </SocialLoginButton>
-          </SocialButtons>
+          <SocialLogin
+            setIsLogin={setIsLogin}
+            setToken={setToken}
+            setErrorMsg={setErrorMsg}
+            isLoading={mutationLogin.isLoading}
+          />
         </Form.Item>
       </Form>
     </Container>
@@ -205,15 +136,6 @@ const ButtonBox = styled.div`
 `;
 const StyledButton = styled(Button)`
   width: 130px;
-`;
-const SocialLoginButton = styled.div`
-  cursor: pointer;
-  margin-top: 10px;
-`;
-const SocialButtons = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
 `;
 const ErrorMessage = styled.p`
   display: flex;
